+++ MatrixSDK/JSONModels/MXMatrixVersions.m
@@ -16,8 +16,19 @@
 
 #import "MXMatrixVersions.h"
 
+const struct MXMatrixClientServerAPIVersionStruct MXMatrixClientServerAPIVersion = {
+    .r0_0_1 = @"r0.0.1",
+    .r0_1_0 = @"r0.1.0",
+    .r0_2_0 = @"r0.2.0",
+    .r0_3_0 = @"r0.3.0",
+    .r0_4_0 = @"r0.4.0",
+    .r0_5_0 = @"r0.5.0",
+};
+
 const struct MXMatrixVersionsFeatureStruct MXMatrixVersionsFeature = {
-    .lazyLoadMembers = @"m.lazy_load_members"
+    .lazyLoadMembers = @"m.lazy_load_members",
+    .requireIdentityServer = @"m.require_identity_server",
+    .idAccessToken = @"m.id_access_token"
 };
 
 @implementation MXMatrixVersions
@@ -35,7 +46,26 @@ const struct MXMatrixVersionsFeatureStruct MXMatrixVersionsFeature = {
 
 - (BOOL)supportLazyLoadMembers
 {
-    return [self.unstableFeatures[MXMatrixVersionsFeature.lazyLoadMembers] boolValue];
+    return [self.versions containsObject:MXMatrixClientServerAPIVersion.r0_5_0]
+        || [self.unstableFeatures[MXMatrixVersionsFeature.lazyLoadMembers] boolValue];
+}
+
+- (BOOL)doesServerRequireIdentityServerParam
+{
+    // YES by default
+    BOOL doesServerRequireIdentityServerParam = YES;
+
+    if (self.unstableFeatures[MXMatrixVersionsFeature.requireIdentityServer])
+    {
+        doesServerRequireIdentityServerParam = [self.unstableFeatures[MXMatrixVersionsFeature.requireIdentityServer] boolValue];
+    }
+
+    return doesServerRequireIdentityServerParam;
+}
+
+- (BOOL)doesServerAcceptIdentityAccessToken
+{
+    return [self.unstableFeatures[MXMatrixVersionsFeature.idAccessToken] boolValue];
 }
 
 @end