+++ CHANGES.rst
@@ -7,6 +7,7 @@ Improvements:
  * MXHTTPClient: Improve M_LIMIT_EXCEEDED error handling: Do not wait to try again if the mentioned delay is too long.
  * MXEventTimeline: The roomEventFilter property is now writable (vector-im/riot-ios#2615).
  * VoIP: Make call start if there is no STUN server.
+ * MXMatrixVersions: Add doesServerRequireIdentityServerParam and doesServerAcceptIdentityAccessToken properties.
 
 Changes in Matrix iOS SDK in 0.13.1 (2019-08-08)
 ===============================================